Move simple components (Footer, NavLink, Filters, TaskInput) into self-c
